Organize website globals and layouts under site
