+++ declaration.tsconfig.json
@@ -9,7 +9,6 @@
   "exclude": [
     "**/*.test.js",
     "test-cases/**/*",
-    "global.data.ts",
-    "layouts/**/*",
+    "site/**/*",
   ]
 }
