+++ site/layouts/docs/navigation-watch.test.js
@@ -4,10 +4,10 @@ import { cp, mkdir, mkdtemp, readFile, rm, stat, writeFile } from 'node:fs/promi
 import { dirname, join, resolve } from 'node:path'
 import { load } from 'cheerio'
 import pino from 'pino'
-import { DomStack } from '../../index.js'
+import { DomStack } from '../../../index.js'
 
 test('heading and index changes refresh shared navigation; body edits leave other pages alone', { timeout: 30_000 }, async t => {
-  const root = resolve(import.meta.dirname, '../..')
+  const root = resolve(import.meta.dirname, '../../..')
   const temp = await mkdtemp(join(root, '.tmp-docs-navigation-'))
   const src = join(temp, 'src')
   const dest = join(temp, 'public')
@@ -16,7 +16,7 @@ test('heading and index changes refresh shared navigation; body edits leave othe
     if (domstack.watching) await domstack.stopWatching()
     await rm(temp, { recursive: true, force: true })
   })
-  for (const file of ['global.data.ts', 'layouts/docs/navigation.js', 'layouts/docs/docs.layout.js']) {
+  for (const file of ['site/globals/global.data.ts', 'site/layouts/docs/navigation.js', 'site/layouts/docs/docs.layout.js']) {
     await mkdir(dirname(join(src, file)), { recursive: true })
     await cp(join(root, file), join(src, file))
   }
